+++ api_tests/entry_api_test.py
@@ -56,35 +56,6 @@ def test_read_write_entries_big_blob_ok(base_url, session, bucket):
     assert resp.headers['x-reduct-last'] == '1'
 
 
-def test_read_no_bucket(base_url, session):
-    """Should return 404 if no bucket found"""
-    resp = session.get(f'{base_url}/b/xxx/entry?ts=100')
-    assert resp.status_code == 404
-    assert 'xxx' in resp.headers["x-reduct-error"]
-
-
-def test_read_no_data(base_url, session, bucket):
-    """Should return 404 if no data for ts"""
-    resp = session.get(f'{base_url}/b/{bucket}/entry?ts=100')
-    assert resp.status_code == 404
-
-
-def test_read_bad_ts(base_url, session, bucket):
-    """Should return 400 if ts is bad"""
-    session.post(f"{base_url}/b/{bucket}/entry?ts=100", data=b"somedata")
-    resp = session.get(f'{base_url}/b/{bucket}/entry?ts=XXXX')
-
-    assert resp.status_code == 422
-    assert "'ts' must be an unix timestamp in microseconds" in resp.headers["x-reduct-error"]
-
-
-def test_read_bad_no_entry(base_url, session, bucket):
-    """Should return 400 if ts is bad"""
-    resp = session.get(f'{base_url}/b/{bucket}/entry')
-    assert resp.status_code == 404
-    assert 'entry' in resp.headers["x-reduct-error"]
-
-
 @requires_env("API_TOKEN")
 def test__read_with_read_token(base_url, session, bucket, token_without_permissions, token_read_bucket,
                                token_write_bucket):
@@ -100,21 +71,6 @@ def test__read_with_read_token(base_url, session, bucket, token_without_permissi
 
     resp = session.get(f'{base_url}/b/{bucket}/entry?ts=1000', headers=auth_headers(token_write_bucket))
     assert resp.status_code == 403
-
-
-def test_write_no_bucket(base_url, session):
-    """Should return 404 if no bucket found"""
-    resp = session.post(f'{base_url}/b/xxx/entry?ts=100')
-    assert resp.status_code == 404
-    assert 'xxx' in resp.headers["x-reduct-error"]
-
-
-def test_write_bad_ts(base_url, session, bucket):
-    """Should return 422 if ts is bad"""
-    resp = session.post(f'{base_url}/b/{bucket}/entry?ts=XXXX')
-    assert resp.status_code == 422
-    assert "'ts' must be an unix timestamp in microseconds" in resp.headers["x-reduct-error"]
-
 
 def test_get_record_ok(base_url, session, bucket):
     """Should return list with timestamps and sizes"""